requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
redis==5.0.7
pypdf==4.3.1
python-dotenv==1.0.1
httpx==0.27.2
tenacity==8.5.0
orjson==3.10.7
